server_mod/python_files/requirements.txt
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
